Added Common proj and tests

+++ Mover/FileMover.cs
@@ -3,7 +3,7 @@ using System.IO;
 
 namespace Mover
 {
-    public class FileMover
+    public class FileMover : IFileMover
     {
         public void Copy(string sourcePath, string targetPath, string fileName)
         {
@@ -27,11 +27,13 @@ namespace Mover
         {
             if (string.IsNullOrEmpty(targetPath)) throw new ArgumentNullException("targetPath is null or empty");
             if (string.IsNullOrEmpty(fileName)) throw new ArgumentNullException("fileName is null or empty");
+            if (!Directory.Exists(targetPath)) throw new DirectoryNotFoundException("targetPath does not exist");
 
             var destinationFile = Path.Combine(targetPath, fileName);
 
             if (!File.Exists(destinationFile)) throw new FileNotFoundException("destinationFile does not exist");
 
+            File.SetAttributes(destinationFile, FileAttributes.Normal);
             File.Delete(destinationFile);
 
             if (File.Exists(destinationFile)) throw new FileNotFoundException("destinationFile exists");
